backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
PyPDF2==3.0.1
python-docx==0.8.11
requests==2.31.0
Werkzeug==2.3.7
python-dateutil==2.8.2
